Tests conversion of AttachmentFile into DiscordImage.
