+++ Tests/Directory.Build.props
@@ -1,0 +1,6 @@
+﻿<Project>
+	<PropertyGroup>
+		<BaseIntermediateOutputPath>..\.obj\Tests</BaseIntermediateOutputPath>
+		<MSBuildProjectExtensionsPath>..\.obj\Tests</MSBuildProjectExtensionsPath>
+	</PropertyGroup>
+</Project>
