+++ Tests/Tests.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>net6.0</TargetFramework>
@@ -8,6 +8,8 @@
     <IsPackable>false</IsPackable>
 
     <RootNamespace>Discord</RootNamespace>
+
+    <BaseOutputPath>..\.bin\Tests</BaseOutputPath>
   </PropertyGroup>
 
   <ItemGroup>
