+++ Tests/appsettings.json
@@ -1,4 +1,4 @@
 {
-  "Token": "<string>",   // A user token
-  "ChannelId": <ulong>   // The channel the unit tests will post to. The user identified by "Token" must have access to this channel.
+  "Token": "",      // string   The user token.
+  "ChannelId": 0    // ulong    The channel the unit tests will post to. The user identified by "Token" must have access to this channel.
 }
